% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry.R
\name{big_a}
\alias{big_a}
\title{Calculate angle between PQ and the X Axis for arbitrary but distinct points P and Q}
\usage{
big_a(px, py, qx, qy)
}
\arguments{
\item{px}{the x-coordinate of point P}

\item{py}{the y-coordinate of point P}

\item{qx}{the x-coordinate of point Q}

\item{qy}{the y-coordinate of point Q}
}
\value{
a numeric angle A between \eqn{[-\pi, \pi]}.
}
\description{
This angle is called \code{A} in
\href{https://www.geeksforgeeks.org/angular-sweep-maximum-points-can-enclosed-circle-given-radius/}{this GeeksForGeeks post}
}
