% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry.R
\name{big_b}
\alias{big_b}
\title{Calculate angle between PC and PQ where C is the centre of the circle on which both P and Q lie}
\usage{
big_b(r, d)
}
\arguments{
\item{r}{the radius of the circle}

\item{d}{the Euclidean distance between points P and Q}
}
\value{
a numeric angle B between \eqn{[0,\pi/2]}.
}
\description{
This angle is called \code{B} in
\href{https://www.geeksforgeeks.org/angular-sweep-maximum-points-can-enclosed-circle-given-radius/}{this GeeksForGeeks post}
}
