% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bikepoints}
\alias{bikepoints}
\title{Santander Cycle docking station locations in London}
\format{
A data frame with 766 rows and 5 variables:
\describe{
\item{id}{TfL API's bike point ID}
\item{common_name}{Text description of the bike point location}
\item{lat}{Latitude of the bike point location}
\item{lon}{Longitude of the bike point location}
\item{no_bikes}{The number of docking stations available at the bike point}
}
}
\usage{
bikepoints
}
\description{
This dataset was produced using \href{https://api.tfl.gov.uk/swagger/ui/index.html?url=/swagger/docs/v1#!/BikePoint/BikePoint_GetAll}{Transport for London's Unified API},
and is contains public sector information licensed under the \href{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/2/}{Open Government Licence v2.0}
}
\keyword{datasets}
