% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{change_projection}
\alias{change_projection}
\title{Change projection of coordinates within a data frame}
\usage{
change_projection(df, xcol, ycol, crs_from, crs_to)
}
\arguments{
\item{df}{the data frame containing the coordinates}

\item{xcol}{the name of the column containing the longitude coordinate}

\item{ycol}{the name of the column containing the latitude coordinate}

\item{crs_from}{the crs projection to convert from}

\item{crs_to}{the crs projection to convert to}
}
\value{
the original data frame with the coordinates converted
}
\description{
Change projection of coordinates within a data frame
}
