% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_matrix}
\alias{dist_matrix}
\title{Calculate distance matrix for \eqn{n} given points}
\usage{
dist_matrix(x)
}
\arguments{
\item{x}{a numeric matrix, data frame or \code{"dist"} object.}
}
\value{
an \eqn{n \times n} distance matrix
}
\description{
Calculate distance matrix for \eqn{n} given points
}
