% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{lonlat_to_utm}
\alias{lonlat_to_utm}
\title{Derive the UTM EPSG coordinate system code for a given location}
\usage{
lonlat_to_utm(lonlat = c(lon = 0, lat = 52))
}
\arguments{
\item{lonlat}{a named numeric vector with \code{lon} and \code{lat} fields}
}
\value{
an integer reflecting the UTM EPSG code
}
\description{
Derive the UTM EPSG coordinate system code for a given location
}
