% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry.R
\name{normalise_angle}
\alias{normalise_angle}
\title{Moves angles less than \eqn{-\pi} or greater than \eqn{\pi} to within this range}
\usage{
normalise_angle(thetas)
}
\arguments{
\item{thetas}{a numeric vector of angles}
}
\value{
a numeric vector of angles within the range \eqn{[-\pi, \pi]}
}
\description{
Moves angles less than \eqn{-\pi} or greater than \eqn{\pi} to within this range
}
