% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry.R
\name{pol2cart}
\alias{pol2cart}
\title{Converts polar coordinates to cartesian}
\usage{
pol2cart(r, theta, origin = c(x = 0, y = 0))
}
\arguments{
\item{r}{vector of radial coordinate(s), the distance from the origin}

\item{theta}{vector of polar angle(s), the counterclockwise angle from the x-axis}

\item{origin}{a named vector (with \code{x} and \code{y} as names) being the location of the origin}
}
\value{
a 2 column matrix with names \code{x} and \code{y} with the cartesian coordinates therein
}
\description{
Converts polar coordinates to cartesian
}
