% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pre_sweep_check}
\alias{pre_sweep_check}
\title{Function to check \link{sweep_latlons} arguments are valid before proceeding}
\usage{
pre_sweep_check(df, xcol, ycol, weight, radius, inc_data)
}
\arguments{
\item{df}{the data frame containing coordinates}

\item{xcol}{the name of the column containing the x coordinate}

\item{ycol}{the name of the column containing the y coordinate}

\item{weight}{name of column containing data with which to weight the points}

\item{radius}{the radius of the circle to sweep with (in metres)}

\item{inc_data}{logical. Whether or not to include the original data alongside
the indices of the points included in each circle}
}
\value{
NULL, if all conditions are met
}
\description{
Function to check \link{sweep_latlons} arguments are valid before proceeding
}
